app/shared/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-cpp-python>=0.3.2
huggingface-hub>=0.23.2
pydantic>=2.10.0
pyyaml>=6.0
